+++ graph/src/data/store/scalar.rs
@@ -320,8 +320,8 @@ impl BigInt {
         BigInt(self.0.pow(&exponent))
     }
 
-    pub fn bits(&self) -> u64 {
-        self.0.bits() as u64
+    pub fn bits(&self) -> usize {
+        self.0.bits()
     }
 }
